requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
redis==5.0.1
celery==5.3.6
pytz==2024.1 